### 2021-08-18 #3628 @whedon
[PRE REVIEW]: Simframe: A Python Framework for Scientific Simulations
**Submitting author:** @stammler (<a href="http://orcid.org/0000-0002-1589-1796">Sebastian Markus Stammler</a>)
**Repository:** <a href="https://github.com/stammler/simframe" target ="_blank">https://github.com/stammler/simframe</a>
**Version:** 0.5.4
