### 2021-08-18 4ef7f13
Added bibliography tag to paper.md
