### 2021-08-18 701da8d
Added JOSS badge to README.md
