### 2021-08-24 #3628 @ketch
I read the paper and I have to say, respectfully, that I am opposed to the promotion of this package.  As the authors state, there are already plenty of existing packages that fulfill this function.  In fact, those packages do it much better than this one.  The package under consideration claims to be helpful for "beginners", but it mainly gives those beginners a set options most of which should never be used by anyone. Beginners would, in my opinion, be much better off using one of the standard existing packages that focus on state-of-the-art methods and have a lot more (numerical and mathematical) functionality.

I'm sure the authors won't be pleased to hear this.  I do respect what they are doing and it is fine to do it.  I just wouldn't recommend that other people use it.

As far as I know, the kind of evaluation I'm doing here is not what is asked for as a reviewer for JOSS, so I think it's best if I don't review this package further.
