### 2021-08-24 #3628 @danielskatz
👋@stammler & @taless474 - based on the comment from @ketch above, let's hold off on reviewing this and do a scope query so editors can make sure it's reasonable for us to review this.  I'll start the process, and it should take a week or two. Because doing this after an editor has been assigned is a little awkward, I'm also going to put a reminder in for 2 weeks for now to make sure we don't drop this.
