### 2021-08-24 #3628 @stammler
Thank you, @danielskatz, for letting us know about the query scope.

And thank you for your sincere comment, @ketch. No offense taken.

Since we think that there might be a slight misunderstanding on the scope of the software we feel the need to clarify that the point of `simframe` is _not_ the integration schemes provided by it, but _mainly_ the framework around it.
With `simframe` it is possible to easily set up data structures for scientific simulations and assign update functions (i.e. how to recalculate one quantity from the rest of the data of the frame) or to define derivatives of the fields, which are then automatically integrated.

The simple integration schemes we provided are just included so the users have some default options to start with. It is a matter of minutes to implement custom integrations schemes as is also described in the documentation. If someone wants to use the more elaborate methods `solve_ivp` or `odeint` of SciPy, this can be done easily. However, in contrast to those methods, `simframe` is less of a black box, and more of a toolbox.

The point of `simframe` is the organization of a scientific simulation or scientific data and the calculation of derived quantities (including the I/O) and the integration is just a convenient extra feature. If the problem does not require integration, it is entirely possible to use `simframe` in a scientific context without integrating anything at all due to its modular structure and its update functionality. `simframe` is, for example, already used by the [DustPy](https://stammler.github.io/dustpy/) project to calculate the evolution of dust in protoplanetary disks and with it there are already at least [eight peer-reviewed publications](https://stammler.github.io/dustpy/A_publications.html) that make use of `simframe`.

We hope this additional information can be helpful for the decision towards a possible JOSS publication.
