### 2021-09-13 #3628 @danielskatz
👋 @openjournals/joss-eics, just a ping that we should try to finish this soon - I put a note in slack to try to move it forward again
