### 2021-11-02 #3882 @whedon
[REVIEW]: Simframe: A Python Framework for Scientific Simulations
**Submitting author:** @stammler (<a href="http://orcid.org/0000-0002-1589-1796">Sebastian Markus Stammler</a>)
**Repository:** <a href="https://github.com/stammler/simframe" target ="_blank">https://github.com/stammler/simframe</a>
**Version:** 1.0.0
