### 2021-11-12 #3882 @taless474
> @taless474: A quick question. We just noticed that whedon did not assign the second reviewer schruste to this review. Is this correct?

@stammler, You are right. But I guess the issue is correct and both @lucaferranti and @schruste are the reviewers. Thank you for noticing it.
