### 2021-11-26 #3882 @schruste
@stammler:
I don't have the rights to set the check marks yet, so I will start with some direct comments:

I think the `simframe` package is an interest package with a well-defined (small) set of features which seem to be however quite useful. Here some further remarks:

* It seems that you have not community guidelines, yet. Perhaps you can add them to your `README.md` ?
* In `paper.md` you have "One one hand". I guess this should be "On the one hand"
* Some basic explanations of the scope of the software that you have in `paper.md` would also be appropriate to put in the `README.md`.
* Especially the ties that you have with `NumPy` should be mentioned early on. Because if I get this correctly you "only" allow for objects/vectors from the `NumPy` universe. In principle I could easily imagine cases where I would like to use your package, but have my own data types for my state vectors, etc.. I think it is not too much of a restriction that you have here, but it would be nice to clarify this early on.
* Have you thought about setting up a small binder? This should be pretty easily set up and would even increase accessibility.
