### 2021-11-27 #3882 @lucaferranti
Hi @stammler , very nice work, here are some comments:

### Paper

- [ ] **comparison to other softwares**: the paper says that there are plenty of ODE solvers packages (which is definitely true), it would be good to mention a few and give a more detailed comparison of simframe to other similar packages, i.e. what features simframe has that other mainstream packages haven't and vice versa.
- [ ] **references**: It is quite strange to me that the paper has no references at all. For example, the integration schemes section could greatly benefit for some references, as they would guide the reader in finding more details about the schemes. Addressing the previous point would also probably lead to some references.

### Documentation

- [ ] Maybe a little surprisingly, I could run all notebooks in the `examples` folder except the first one, it gets stuck at the fourth last cell `sim_cont.run()` and then the kernel dies. Is that example somehow particularly computationally expensive? If it is, it would be good to mention it explicitly in the notebook. I'll retry this later to verify it wasn't a strange glitch on my side.
- [ ] (minor) It would be nice to add in each example in the docs a "download the notebook" button and/or a "try it on binder" button if you setup binder as suggested by the other reviewer.
- [ ] **automated tests**: I would strongly suggest to setup a github actions (or travis, or your favourite CI tool) to automatically run the tests whenever someone pushes (to master or through PR), this is mainly to make sure nobody accidentally breaks things while contributing. Github should have plenty of workflow-templates for python packages, you can find those from the `actions` tab. It would also be nice to have some coverage tool such as codecov or coveralls
- [ ] **contribution guidelines**: As pointed out by the other reviewer, contribution guidelines are missing. Github has a cheatsheet of "good social practices" under the repo `insights > community`. There you can find a few tips to write contributions guidelines.
