### 2021-12-01 ad90dd9
Adding assert=plain to automated test
