### 2021-12-01 dd1c39b
Removing assert=plain from automated test
