### 2021-12-02 d8a4271
Modification to writer test
