### 2021-12-03 f95c0c9
Adding fixes to automated tests
