+++ tests/io/test_hdf5writer.py
@@ -20,6 +20,8 @@ def test_hdf5writer_skip():
 
 def test_hdf5writer_strings():
     string = "test"
+    # When read from HDF5 the string will be a byte literal
+    string_cmpr = string.encode()
     f = Frame()
     f.addfield("s", string)
     f.t = string
@@ -27,15 +29,15 @@ def test_hdf5writer_strings():
     f.writeoutput(0)
     f.writeoutput(1)
     data0000 = f.writer.read.output(0)
-    assert data0000.s[0] == string
-    assert data0000.t == string
+    assert data0000.s[0] == string_cmpr
+    assert data0000.t == string_cmpr
     s = f.writer.read.sequence("s")
-    assert np.all(s == [string, string])
+    assert np.all(s == [string_cmpr, string_cmpr])
     t = f.writer.read.sequence("t")
-    assert np.all(t == [string, string])
+    assert np.all(t == [string_cmpr, string_cmpr])
     data = f.writer.read.all()
-    assert np.all(data.s == [string, string])
-    assert np.all(data.s == [string, string])
+    assert np.all(data.s == [string_cmpr, string_cmpr])
+    assert np.all(data.s == [string_cmpr, string_cmpr])
     shutil.rmtree(f.writer.datadir)
 
 
