+++ paper/joss/paper.bib
@@ -1,0 +1,17 @@
+@article{harris2020Natur.585..357H,
+  author        = {{Harris}, Charles R. and {Millman}, K. Jarrod and {van der Walt}, St{\'e}fan J. and {Gommers}, Ralf and {Virtanen}, Pauli and {Cournapeau}, David and {Wieser}, Eric and {Taylor}, Julian and {Berg}, Sebastian and {Smith}, Nathaniel J. and {Kern}, Robert and {Picus}, Matti and {Hoyer}, Stephan and {van Kerkwijk}, Marten H. and {Brett}, Matthew and {Haldane}, Allan and {del R{\'\i}o}, Jaime Fern{\'a}ndez and {Wiebe}, Mark and {Peterson}, Pearu and {G{\'e}rard-Marchant}, Pierre and {Sheppard}, Kevin and {Reddy}, Tyler and {Weckesser}, Warren and {Abbasi}, Hameer and {Gohlke}, Christoph and {Oliphant}, Travis E.},
+  title         = {{Array programming with NumPy}},
+  journal       = {\nat},
+  keywords      = {Computer Science - Mathematical Software, Statistics - Computation},
+  year          = 2020,
+  month         = sep,
+  volume        = {585},
+  number        = {7825},
+  pages         = {357-362},
+  doi           = {10.1038/s41586-020-2649-2},
+  archiveprefix = {arXiv},
+  eprint        = {2006.10256},
+  primaryclass  = {cs.MS},
+  adsurl        = {https://ui.adsabs.harvard.edu/abs/2020Natur.585..357H},
+  adsnote       = {Provided by the SAO/NASA Astrophysics Data System}
+}
