+++ paper/joss/paper.md
@@ -23,7 +23,7 @@ bibliography: paper.bib
 
 `Simframe` is a Python framework to facilitate scientific simulations. It can be used to easily integrate differential equations.
 
-Conceptually, upon initialization `Simframe` is an empty frame that can be filled with `Field`s containing the data. `Field`s are derived from `numpy.ndarray`s [@harris2020Natur.585..357H], but with extended functionality. The user can then specify differential equations to those data fields and can set up an integrator which is integrating those fields according the given differential equations.
+Conceptually, upon initialization `Simframe` is an empty frame that can be filled with `Field`s containing the data. `Field`s are derived from `numpy.ndarray`s [@Harris:2020], but with extended functionality. The user can then specify differential equations to those data fields and can set up an integrator which is integrating those fields according the given differential equations.
 
 Fields that should not be integrated but are still required for the model can be assigned an update function, according to which they will be updated once per integration step.
 
