+++ .github/CONTRIBUTING.md
@@ -1,0 +1,17 @@
+# How to contribute?
+
+We invite everyone to contribute to the further development of `simframe`. Either by directly contributing to the code, by reporting bugs, or by requesting features.
+
+## Reporting bugs
+
+If you encountered a bug in `simframe`, please about a new [bug report issue](https://github.com/stammler/simframe/issues/new?template=bug_report.md&title=[BUG]+Descriptive+title+of+the+bug+report) and describe the bug, the expected behavior, and steps how to reproduce the bug.
+
+## Requesting features
+
+If you have an idea of a feature, that is missing in `simframe`, of if you want to suggest and improvement, please open a new [feature request issue](https://github.com/stammler/simframe/issues/new?template=feature_request.md&title=[FEATURE]+Descriptive+title+of+the+feature+request).
+
+## Contributing code
+
+To contribute code please open a new [pull request]() and describe the changes to the software your pull request introduces.
+
+Please note, that we want to achieve a code coverage of 100%. Any addition to the software must therefore also come with unit tests. Additional features must also be described in the documentation.
