### 2021-12-03 5c80f84
Adding link to contribution guidelines in README.md
