+++ paper/joss/paper.bib
@@ -9,3 +9,15 @@
   title         = {{Array programming with NumPy}},
   year          = 2020
 }
+
+@article{Virtanen:2020,
+  url           = {https://ui.adsabs.harvard.edu/abs/2020NatMe..17..261V},
+  archiveprefix = {arXiv},
+  author        = {{Virtanen}, Pauli and {Gommers}, Ralf and {Oliphant}, Travis E. and {Haberland}, Matt and {Reddy}, Tyler and {Cournapeau}, David and {Burovski}, Evgeni and {Peterson}, Pearu and {Weckesser}, Warren and {Bright}, Jonathan and {van der Walt}, St{\'e}fan J. and {Brett}, Matthew and {Wilson}, Joshua and {Millman}, K. Jarrod and {Mayorov}, Nikolay and {Nelson}, Andrew R.~J. and {Jones}, Eric and {Kern}, Robert and {Larson}, Eric and {Carey}, C.~J. and {Polat}, {\.I}lhan and {Feng}, Yu and {Moore}, Eric W. and {VanderPlas}, Jake and {Laxalde}, Denis and {Perktold}, Josef and {Cimrman}, Robert and {Henriksen}, Ian and {Quintero}, E.~A. and {Harris}, Charles R. and {Archibald}, Anne M. and {Ribeiro}, Ant{\^o}nio H. and {Pedregosa}, Fabian and {van Mulbregt}, Paul and {SciPy 1. 0 Contributors}},
+  eprint        = {1907.10121},
+  journal       = {Nature},
+  keywords      = {Computer Science - Mathematical Software, Computer Science - Data Structures and Algorithms, Computer Science - Software Engineering, Physics - Computational Physics},
+  month         = feb,
+  title         = {{SciPy 1.0: fundamental algorithms for scientific computing in Python}},
+  year          = 2020
+}
