### 2021-12-03 43fd44b
Adding matplotlib to requirements for binder
