+++ requirements.txt
@@ -1,4 +1,5 @@
 dill
 h5py
+matplotlib
 numpy
 scipy
