+++ .github/PULL_REQUEST_TEMPLATE/pull_request_template.md
@@ -1,0 +1,13 @@
+Please have a look the [contribution guidelines](https://github.com/stammler/simframe/blob/master/.github/CONTRIBUTING.md).
+
+# Reference issue
+
+Please reference the issue this pull request is referring to, if applicable.
+
+# Describe the changes
+
+Please describe the changes or fixes this pull request introduces into the project.
+
+# Additional information
+
+Please provide additional information that is helpful for this pull request.
