+++ .github/PULL_REQUEST_TEMPLATE.md
@@ -1,13 +1,15 @@
-Please have a look the [contribution guidelines](https://github.com/stammler/simframe/blob/master/.github/CONTRIBUTING.md).
+# Pull Requests
 
-# Reference issue
+Please have a look the [contributing guidelines](https://github.com/stammler/simframe/blob/master/.github/CONTRIBUTING.md).
+
+## Reference issue
 
 Please reference the issue this pull request is referring to, if applicable.
 
-# Describe the changes
+## Describe the changes
 
 Please describe the changes or fixes this pull request introduces into the project.
 
-# Additional information
+## Additional information
 
 Please provide additional information that is helpful for this pull request.
