### 2021-12-07 daa9ad3
Changes to contributing guidelines
