### 2021-12-07 e7191f3
Changes to bibliography
