+++ paper/joss/paper.bib
@@ -1,3 +1,11 @@
+@book{Hairer:1993,
+  author    = {{Hairer}, Ernst and {Nørsett} Syvert Paul and {Wanner}, Gerhard},
+  title     = {Solving Ordinary Differential Equations I},
+  publisher = {Springer, Berlin, Heidelberg},
+  year      = {1993},
+  series    = {Springer Series in Computational Mathematics}
+}
+
 @article{Harris:2020,
   url           = {https://ui.adsabs.harvard.edu/abs/2020Natur.585..357H},
   archiveprefix = {arXiv},
