### 2021-12-07 51f1dbd
Adding citations to paper.md
