+++ paper/joss/paper.md
@@ -21,11 +21,11 @@ bibliography: paper.bib
 
 # Summary
 
-`Simframe` is a Python framework to facilitate scientific simulations. It can be used to easily integrate differential equations.
+`Simframe` is a Python framework to facilitate scientific simulations. The scope of the software is to provide a framework which can hold data fields, which can be used to integrate differential equations, and which can read and write data files.
 
-Conceptually, upon initialization `Simframe` is an empty frame that can be filled with `Field`s containing the data. `Field`s are derived from `numpy.ndarray`s [@Harris:2020], but with extended functionality. The user can then specify differential equations to those data fields and can set up an integrator which is integrating those fields according the given differential equations.
+Conceptually, upon initialization `Simframe` is an empty frame that can be filled with `Field`s containing data. `Field`s are derived from `numpy.ndarray`s [@Harris:2020], but with extended functionality. The user can then specify differential equations to those data fields and can set up an integrator which is integrating those fields according the given differential equations. Therefore, `Simframe` can only work with data, that can be stored in a `NumPy` array.
 
-Fields that should not be integrated but are still required for the model can be assigned an update function, according to which they will be updated once per integration step.
+Data fields that should not be integrated themselves, but are still required for the model, can have an update function assigned to them, according to which they will be updated once per integration step.
 
 `Simframe` contains a number of integration schemes of different orders, both for explicit and implicit integration. Furthermore, `Simframe` includes methods to read and write output files.
 
@@ -35,11 +35,11 @@ Due to its modular structure, `Simframe` can be extended at will, for example, b
 
 Solving differential equations is part of the daily work of scientists. `Simframe` facilitates this by providing the infrastructure: Data structures, integration schemes, and methods to write and read output files.
 
-On one hand, `Simframe` can be used to quickly solve small scientific problems, and, on the other hand, can be easily extended to larger projects due to its versatility and modular structure.
+On one hand, `Simframe` can be used to quickly solve small scientific problems, and, on the other hand, it can be easily extended to larger projects due to its versatility and modular structure.
 
 Furthermore, `Simframe` is ideal for beginners without programming experience who are taking their first steps in solving differential equations. It can therefore be used to design lectures or practical courses at schools and universities, as it allows students to concentrate on the essentials without having to write larger programs on their own.
 
-While plenty of ODE solver packages exist, `Simframe` offers a flexible framework to define, group, and describe data, define how it is updated, use existing integrators or define new ones, and to handle writing of data or serializing the entire simulation object, all in one modular package.
+While plenty of ODE solver packages exist for Python, like `solve_ivp` or `odeint` in `SciPy`'s `integrate` module. These functions, however, do not provide data structures, nor input/output capabilities. `Simframe` offers a flexible framework to define, group, and describe data, define how it is updated, use existing integrators or define new ones, and to handle writing of data or serializing the entire simulation object, all in one modular package.
 
 # Features
 
@@ -49,7 +49,7 @@ The data fields of `Simframe` are subclassed `NumPy` `ndarray`s. The full `NumPy
 
 ## Integration schemes
 
-`Simframe` includes a number of integration schemes by default [@Hairer:1993]. All of the implemented schemes are Runge-Kutta methods of different orders. Some of the methods are adaptive, i.e., they are embedded Runge-Kutta methods, that return an optimal step size for the integration variable, such that the desired accuracy is achieved. The implicit methods require a matrix inversion that is either done directly by `NumPy` or by using the GMRES solver provided by `SciPy` [@Virtanen:2020].
[…]
