+++ paper/joss/paper.bib
@@ -29,3 +29,20 @@
   title         = {{SciPy 1.0: fundamental algorithms for scientific computing in Python}},
   year          = 2020
 }
+
+@book{Collette:2014,
+  keywords  = {python, hdf5},
+  year      = {2013},
+  publisher = {O'Reilly},
+  title     = {Python and HDF5},
+  author    = {{Collette}, Andrew}
+}
+
+@misc{McKerns:2012,
+  title         = {Building a Framework for Predictive Science},
+  author        = {{McKerns}, Michael M. and {Strand}, Leif and {Sullivan}, Tim and {Fang}, Alta and {Aivazis}, Michael A. G.},
+  year          = {2012},
+  eprint        = {1202.1056},
+  archiveprefix = {arXiv},
+  primaryclass  = {cs.MS}
+}
