### 2021-12-08 #3882 @stammler
Dear  @schruste  and @lucaferranti!

Thank you for your great comments!

We added contribution guidelines, templates for bug report and feature request issues, and a template for pull requests.
We furthermore added a small explanation about the scope of the software in the README.md.

We changed the license to the more permissive BSD 3-Clause License.

We clarified in the paper and the README.md that `simframe` can only handle data, that can be stored within NumPy arrays. We furthermore added references to the used packages and the integration schemes.

All example notebooks of the documentation have now a button to launch them on binder.
Since the first notebook works fine on Binder and on Readthedocs for building the documentation, we assume that your reported problem, @lucaferranti, is local on your setup for now. Please let us know if the problem persists.

We added automated unit tests to Github Actions on pull requests and merges to master.

We would like to refrain from directly comparing `simframe` to existing ODE solvers. The point of `simframe` is not to provide solvers, but to provide an infrastructure to run scientific simulations including data structures and input/output methods. The simple integration schemes that we provide by default are only a bonus, such that `simframe` can be used as is. Existing solvers like `solve_ivp` or `odeint`, on the other hand, can be used as integration scheme within `simframe`. We clarified this in the paper.

Thanks again!
And please let us know if you have more comments.
