### 2021-12-08 #3882 @lucaferranti
hi @stammler 👋 ,

I'll have a closer look today or tomorrow, overall it looks very good!

One quick small comment: I noticed your workflow runs only on ubuntu, have you considering running it also on windows and macos? I think it would make the testing more robust to possible small OS dependent bugs
