### 2021-12-08 4f05e8e
Adding Windows and MacOS to unit tests
