### 2021-12-08 9ff9f33
Remove condition in workflow
