### 2021-12-08 bf41e86
Bugfix to writer extension test
