### 2021-12-08 f6000ea
Avoiding division by zero in progress bar ETA
