### 2021-12-08 #3882 @stammler
Thanks!
We added Windows and MacOS to the workflow.
