### 2021-12-13 #3882 @lucaferranti
@stammler @taless474 after the latest changes I think this submission is good to go. There might be a missing doi in the reference, but after that is checked I think the paper is ready for publication.
