+++ paper/joss/paper.bib
@@ -3,7 +3,8 @@
   title     = {Solving Ordinary Differential Equations I},
   publisher = {Springer, Berlin, Heidelberg},
   year      = {1993},
-  series    = {Springer Series in Computational Mathematics}
+  series    = {Springer Series in Computational Mathematics},
+  doi       = {10.1007/978-3-540-78862-1}
 }
 
 @article{Harris:2020,
@@ -13,9 +14,10 @@
   eprint        = {2006.10256},
   journal       = {Nature},
   keywords      = {Computer Science - Mathematical Software, Statistics - Computation},
-  month         = mar,
+  month         = {mar},
   title         = {{Array programming with NumPy}},
-  year          = 2020
+  year          = {2020},
+  doi           = {10.1038/s41586-020-2649-2}
 }
 
 @article{Virtanen:2020,
@@ -25,9 +27,10 @@
   eprint        = {1907.10121},
   journal       = {Nature},
   keywords      = {Computer Science - Mathematical Software, Computer Science - Data Structures and Algorithms, Computer Science - Software Engineering, Physics - Computational Physics},
-  month         = feb,
+  month         = {feb},
   title         = {{SciPy 1.0: fundamental algorithms for scientific computing in Python}},
-  year          = 2020
+  year          = {2020},
+  doi           = {10.1038/s41592-019-0686-2}
 }
 
 @book{Collette:2014,
@@ -44,5 +47,6 @@
   year          = {2012},
   eprint        = {1202.1056},
   archiveprefix = {arXiv},
-  primaryclass  = {cs.MS}
+  primaryclass  = {cs.MS},
+  doi           = {10.25080/majora-ebaa42b7-00d}
 }
