### 2021-12-13 4411f59
Update of Acknowledgements in paper.md
