+++ paper/joss/paper.md
@@ -11,10 +11,12 @@ authors:
     affiliation: 1
   - name: Tilman Birnstiel
     orcid: 0000-0002-1899-8783
-    affiliation: 1
+    affiliation: 1, 2
 affiliations:
   - name: University Observatory, Faculty of Physics, Ludwig-Maximilians-Universität München, Scheinerstr. 1, 81679 Munich, Germany
     index: 1
+  - name: Exzellenzcluster ORIGINS, Boltzmannstr. 2, D-85748 Garching, Germany
+    index: 2
 date: 11 August 2021
 bibliography: paper.bib
 ---
@@ -84,6 +86,6 @@ If configured by the user, `Simframe` is writing dump files, from which the simu
 
 # Acknowledgements
 
-The authors acknowledge funding from the European Research Council (ERC) under the European Union's Horizon 2020 research and innovation programme under grant agreement No 714769 and funding by the Deutsche Forschungsgemeinschaft (DFG, German Research Foundation) - under Ref no. FOR 2634 - 325594231 and under Germany's Excellence Strategy - EXC-2094 - 390783311.
+The authors acknowledge funding from the European Research Council (ERC) under the European Union's Horizon 2020 research and innovation programme under grant agreement No 714769 and funding by the Deutsche Forschungsgemeinschaft (DFG, German Research Foundation) under grants 361140270, 325594231, and Germany's Excellence Strategy - EXC-2094 - 390783311.
 
 # References
