### 2021-12-15 #3882 @stammler
A quick question about the procedure:
Is now the point where we should upload it to Zenodo and provide a DOI or is there something else that has to be done first?
