### 2021-12-15 #3882 @taless474
> A quick question about the procedure: Is now the point where we should upload it to Zenodo and provide a DOI or is there something else that has to be done first?

Hi @stammler, yes please go ahead and upload it. You can refer to links in [here](https://joss.readthedocs.io/en/latest/editing.html).
@lucaferranti and @schruste, thank you for your revisions, good job 👍
