### 2021-12-16 #3882 @stammler
@taless474: the DOI is [10.5281/zenodo.5785575](https://doi.org/10.5281/zenodo.5785575 ). The version number is 1.0.0

@lucaferranti & @schruste: Thank you for your great comments!
