### 2022-01-11 #3882 @stammler
Not sure if this matters. But the latest version is 1.0.0, while the top comment of the issue still says 0.5.4
