### 2022-01-11 #3882 @danielskatz
@stammler - I'm the next step in the process. I'll proofread it in the next couple of hours, and either let you know if any changes need to be made, or will proceed to publication.
